+++ core/docs/changelog.md
@@ -6,10 +6,12 @@
 * Build the interactive CLI & Javascript REPL image with `build-and-run.sh`
 * Split `build_and_run.sh` into two scripts: `build.sh` and `run-internal-testsuite.sh`
 * Added a `launch-interactive.sh` script for running Kurtosis Interactive
+* Publish the Javascript REPL image to `kurtosistech/javascript-interactive-repl`
 
 ### Fixes
 * Fix bug with not checking enclave creation error in the interactive CLI
 * The CircleCI artifact-publishing now uses the same constants as all the other scripts in the `scripts` directory
+* The `launch-interactive` script will now appropriately use whichever version of the Javascript repl that you're working on
 
 # 1.18.4
 ### Changes
@@ -27,7 +29,7 @@
 
 # 1.18.2
 ### Changes
-* Switched some less-important log levels (e.g. "Startin API container..." from INFO -> DEBUG)
+* Switched some less-important log levels (e.g. "Starting API container..." from INFO -> DEBUG)
 
 ### Features
 * Added a watermark with support information that displays on every run