+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.18.5
 ### Changes
 * The interactive CLI now requires an API container image version